+++ internal/rest/controllers/user.go
@@ -1,6 +1,7 @@
 package controllers
 
 import (
+	"foosball/internal/rest/handlers"
 	"foosball/internal/rest/helpers"
 	"foosball/internal/user"
 	"net/http"
@@ -8,9 +9,9 @@ import (
 	"github.com/labstack/echo/v4"
 )
 
-func (h *Handlers) DeleteUser(c echo.Context) error {
+func (h *Handlers) DeleteUser(c handlers.AuthenticatedContext) error {
 	type deleteUserRequest struct {
-		ID uint `param:"userId" validate:"required"`
+		ID uint `param:"userId" validate:"required, gt=0"`
 	}
 
 	ctx := c.Request().Context()
@@ -18,6 +19,10 @@ func (h *Handlers) DeleteUser(c echo.Context) error {
 	req, err := helpers.Bind[deleteUserRequest](c)
 	if err != nil {
 		return echo.ErrBadRequest
+	}
+
+	if req.ID != c.Claims.UserID {
+		return echo.ErrUnauthorized
 	}
 
 	if err := h.userService.DeleteUser(ctx, req.ID); err != nil {
@@ -28,9 +33,9 @@ func (h *Handlers) DeleteUser(c echo.Context) error {
 	return c.NoContent(http.StatusOK)
 }
 
-func (h *Handlers) GetUsersInOrganization(c echo.Context) error {
+func (h *Handlers) GetUsersInOrganization(c handlers.AuthenticatedContext) error {
 	type getUsersInOrgRequest struct {
-		OrgId uint `json:"orgId" validate:"required"`
+		OrgId uint `json:"orgId" validate:"required, gt=0"`
 	}
 
 	type getUsersInOrgResponse struct {
@@ -42,6 +47,10 @@ func (h *Handlers) GetUsersInOrganization(c echo.Context) error {
 	req, err := helpers.Bind[getUsersInOrgRequest](c)
 	if err != nil {
 		return echo.ErrBadRequest
+	}
+
+	if req.OrgId != c.Claims.OrganizationID {
+		return echo.ErrUnauthorized
 	}
 
 	users, err := h.userService.GetUsersInOrganization(ctx, req.OrgId)
@@ -57,10 +66,10 @@ func (h *Handlers) GetUsersInOrganization(c echo.Context) error {
 	return c.JSON(http.StatusOK, reponse)
 }
 
-func (h *Handlers) RemoveUserFromOrganization(c echo.Context) error {
+func (h *Handlers) RemoveUserFromOrganization(c handlers.AuthenticatedContext) error {
 	type removeUserFromOrgRequest struct {
-		OrgId  uint `param:"orgId" validate:"required"`
-		UserId uint `param:"userId" validate:"required"`
+		OrgId  uint `param:"orgId" validate:"required, gt=0"`
+		UserId uint `param:"userId" validate:"required, gt=0"`
 	}
 
 	ctx := c.Request().Context()
@@ -68,6 +77,10 @@ func (h *Handlers) RemoveUserFromOrganization(c echo.Context) error {
 	req, err := helpers.Bind[removeUserFromOrgRequest](c)
 	if err != nil {
 		return echo.ErrBadRequest
+	}
+
+	if !(c.Claims.Admin && req.OrgId == c.Claims.OrganizationID) || req.UserId == c.Claims.UserID {
+		return echo.ErrUnauthorized
 	}
 
 	users, err := h.userService.GetUsers(ctx, []uint{req.UserId})
@@ -89,10 +102,10 @@ func (h *Handlers) RemoveUserFromOrganization(c echo.Context) error {
 	return c.NoContent(http.StatusOK)
 }
 
-func (h *Handlers) SetUserAsAdmin(c echo.Context) error {
+func (h *Handlers) SetUserAsAdmin(c handlers.AuthenticatedContext) error {
 	type setUserAsAdminRequest struct {
-		OrgId  uint `param:"orgId" validate:"required"`
-		UserId uint `param:"userId" validate:"required"`
+		OrgId  uint `param:"orgId" validate:"required, gt=0"`
+		UserId uint `param:"userId" validate:"required, gt=0"`
 	}
 
 	ctx := c.Request().Context()
@@ -100,6 +113,10 @@ func (h *Handlers) SetUserAsAdmin(c echo.Context) error {
 	req, err := helpers.Bind[setUserAsAdminRequest](c)
 	if err != nil {
 		return echo.ErrBadRequest
+	}
+
+	if !(c.Claims.Admin && req.OrgId == c.Claims.OrganizationID) {
+		return echo.ErrUnauthorized
 	}
 
 	users, err := h.userService.GetUsers(ctx, []uint{req.UserId})
@@ -121,10 +138,10 @@ func (h *Handlers) SetUserAsAdmin(c echo.Context) error {
 	return c.NoContent(http.StatusOK)
 }
 
-func (h *Handlers) RemoveAdminFromUser(c echo.Context) error {
+func (h *Handlers) RemoveAdminFromUser(c handlers.AuthenticatedContext) error {
 	type removeAdminFromUserRequest struct {
-		OrgId  uint `param:"orgId" validate:"required"`
-		UserId uint `param:"userId" validate:"required"`
+		OrgId  uint `param:"orgId" validate:"required, gt=0"`
+		UserId uint `param:"userId" validate:"required, gt=0"`
 	}
 
 	ctx := c.Request().Context()
@@ -132,6 +149,10 @@ func (h *Handlers) RemoveAdminFromUser(c echo.Context) error {
 	req, err := helpers.Bind[removeAdminFromUserRequest](c)
 	if err != nil {
 		return echo.ErrBadRequest
+	}
+
+	if !(c.Claims.Admin && req.OrgId == c.Claims.OrganizationID) {
+		return echo.ErrUnauthorized
 	}
 
 	users, err := h.userService.GetUsers(ctx, []uint{req.UserId})
